pyproject.toml
[tool.poetry]
name = "lander"
version = "0.1.0"
description = ""
authors = ["Mai1902 <nguyendacm@allegheny.edu>"]

[tool.poetry.dependencies]
python =  ">=3.9,<3.9.7 || >3.9.7,<4.0"
bs4 = "^0.0.1"
pandas = "^1.5.0"
PyPDF2 = "^2.11.0"
spacy = "^3.4.1"
sklearn = "^0.0"
resume-parser = "^0.8.4"
gensim = "^4.2.0"
rake-nltk = "^1.0.6"
en-core-web-sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl"}
streamlit = "^1.18.1"
Pillow = "^9.4.0"


[tool.poetry.dev-dependencies]
requests = "^2.28.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.core.masonry.api"
